Wave 30: 2258 -> 2448 (+190), bytes 23.43% -> 24.95%
Five agents at a concurrency cap of three, on the address-locality axis.
193 of 198 targets verified by exit code, 190 newly promoted; both builds
reproduce 14dd0b22c894865867aff89e8116b2dffae25605.

MEASURED: BLOCK DENSITY IS NOT A COST PREDICTOR, AND THE THEORY WAS LOOKING
AT THE WRONG VOCABULARY. W30-A ran three blocks at density 9/10/10 -- below
anything previously tested -- and went 45 for 45, mean 1.24 try_match
attempts (37 functions at one attempt, seven at two, one at four).

The mechanism the theory rests on is REAL and large: 29 of 45 functions
(64%) needed a global, struct member or prototype that no already-matched
neighbour in their block had. It just does not cost anything -- 24 of those
29 still fell in one attempt, because a global's width and signedness read
straight off the target's own load instructions in the same pass that reads
the body. What actually cost attempts was CALLEE PROTOTYPES (a missing third
parameter on sub_08080E74 and sub_08044B28), and callees are not
address-local: those two live nowhere near the 0x08039 block, so their
availability is uncorrelated with how promoted it is.

One case inverted the prediction outright. sub_080321F0 was slow because
gUnknown_0849B018->unk06 was ALREADY declared -- wrongly but byte-neutrally,
by a matched function 0x2000 bytes away. Existing vocabulary was the
obstacle.

So: stop screening on block density; density 9 is a fine place to work.
W30-A's proposed replacement, recorded in docs/agbcc-codegen.md, is to
prefer blocks whose members' `bl` targets are already declared in
include/unknown-functions.h, and to treat a promoted-but-undeclared callee
(three in that batch: sub_080337D8, sub_08039BB4, sub_08044144) as free
rather than as a gap.

A CALLER'S `asrs` IS A CAST AT A USE, NOT A RETURN TYPE -- and this is a
refinement of the proto_check discriminator, not an exception to it.
sub_080157D0 / sub_08015820 were declared `s16` this wave on the reasoning
that every caller re-narrows with `lsls #0x10; asrs #0x10` and then compares
SIGNED. That is a genuine call-site observation and it is still the wrong
conclusion. Both bodies are one `return tbl[i].field;` compiled to
`ldrh r0, [r0, #60]` flowing into `bx lr` -- an UNSIGNED load with no
re-narrowing. `s16` forces `ldrsh`, which needs its offset in a register,
and costs +4 bytes on a 36-byte function (measured: 77.8%). The callers'
arithmetic shift is an explicit `(s16)` cast in the CALLER's source, exactly
what the copy-then-narrow rule says a narrowing after a `bl` means when the
value is used afterwards, and a signed compare downstream constrains the
caller's local rather than the callee's return.

The rule that survives: caller-side facts beat the definition, but `ldrh`
vs `ldrsh` in the callee's own body is itself a hard fact and it outranks
call-site SHAPE. Reverted to `u16`; corrected in place in
include/unknown-functions.h so no wave re-derives it.

The other two proto_check conflicts settled the other way, toward the
callers, and both were byte-neutral: sub_08034534's third and fourth
parameters are `u8` (sub_08041E48 narrows each with `lsls #0x18; lsrs #0x18`
before the bl -- 8 bytes an `int` parameter does not emit; the previous
`int, int` rested only on a caller passing literal zeros, byte-identical at
any width).

proto_check COMPARES DECLARATION TEXT AND DOES NOT RESOLVE TYPEDEFS.
Its fourth "mismatch" was `u8 sub_080266DC(u8)` against a `bool8` definition.
`bool8` IS `u8` (include/gba/types.h:27), so there was nothing to settle and
no caller could ever have been affected. Spelled the declaration `bool8` to
clear it. Worth fixing in the tool: a false positive here costs a real
investigation at exactly the moment a wave is trying to tell real conflicts
apart.

TORN READS CUT BOTH WAYS, AND THAT IS THE ORCHESTRATION LESSON.
A per-agent sweep run while other agents were still editing shared headers
produced 27 failures all pointing at one unknown-globals.h line; every one
passed on a plain re-run. Adding an automatic retry on the
`conflicting types|redefinition` signature made later per-agent sweeps
usable. But the same noise HID two real failures: sub_0801C4D4 and
sub_0801C640 were in that batch and genuinely fail at 86.1% and 88.3%. Only
the final from-scratch sweep of all 198, run after every agent was idle,
distinguished them. Per-agent sweeps are triage; the gate is the full sweep
at the end.

REMOVE A PARKED ENTRY WHEN A LATER AGENT MATCHES IT. W30-C parked
sub_0807610C and sub_080763C0 with its remaining diffs; W30-E then matched
both. Left in place they would have suppressed two matched functions from
every future pre-flight. Deleted at integration.

Handing a follow-up agent the PRE-READ DIFF closed both: sub_0807610C was
`orrs r1,r0` vs `orrs r0,r1` (the ROM modifies the loaded value in place,
the draft accumulated into the constant) and sub_080763C0 was gPal and
gUnknown_081D22A4 swapped in the literal pool, fixed through source
reference order rather than by chasing the pool, per the wave-17 allocation
rule. W30-E went 14 for 14 on a batch built entirely from other agents'
leftovers.

data/parked.json IS CRLF AS WELL AS indent=1. The wave-26 note records only
the indent half. Writing it with LF re-serialises all 379 lines and buries a
five-entry addition in a 425/379 diff; with both right it is 44 insertions
and zero deletions.

Parked five with exact diffs and ruled-out axes. sub_0801C4D4 and
sub_0801C640 are the same local-alloc tie-break in one block, each with
seven axes ruled out by W30-B and a shared note that one probe at agbcc's
allocation priority would likely unblock both. sub_08087514 is NOT an
allocation problem: the ROM builds 187 as `(5 - 2) + 184` at run time, and
gcc does not leave arithmetic on two literals unfolded, so the `2` is a
variable holding 2 on that path -- a source-recovery job with a concrete
target. sub_08073930 carries a pointer to the prior art already in
unknown-globals.h that a fresh attempt would otherwise re-derive wrongly.

+++ src/decomp/c_08028874.c
@@ -7,14 +7,13 @@
  * sub_08028874 @ 0x08028874
  */
 
-/* Promoted from assembly; each function below is byte-for-byte
- * identical to the original. Order is address order and must
- * stay that way -- the linker places this file's .text as one
- * contiguous block at 0x08028874.
- * sub_08028874 @ 0x08028874
- */
-
-void sub_08028874(int index, u8 value)
+/* `value` is `int`, RETYPED IN WAVE 30 (W30-B) from `u8`. Nothing in this body
+ * can tell the two apart -- the value's only use is a `strb` into a u8 member,
+ * which is byte-identical at every width -- so the evidence is entirely on the
+ * caller side. sub_08028894 passes its own raw `int` parameter here with no
+ * narrowing, where a `u8` parameter makes agbcc emit `lsls #0x18; lsrs #0x18`
+ * in front of the `bl`. */
+void sub_08028874(int index, int value)
 {
     gUnknown_08499598[index].unk14 = gUnknown_03004080;
     gUnknown_08499598[index].unk13 = value;
